Add new recipes
I did this project in November 2022, but I stopped studying programming in 2023 to focus on my high school studies. Now I'm a college student on Computer Science, and I returned the studies on The Odin Project. Because of that, I decided to add more recipes to this project instead of making another one from the start.